Refactor the EdpSimulator to match the style of other simulators.
The key change is defining a signle method .process() to cover all works
the simulator does.
