+++ src/main/kotlin/org/wfanet/measurement/loadtest/dataprovider/EdpSimulatorFlags.kt
@@ -1,0 +1,103 @@
+// Copyright 2021 The Cross-Media Measurement Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package org.wfanet.measurement.loadtest.dataprovider
+
+import java.io.File
+import java.time.Duration
+import kotlin.properties.Delegates
+import org.wfanet.measurement.common.grpc.TlsFlags
+import org.wfanet.measurement.loadtest.KingdomPublicApiFlags
+import org.wfanet.measurement.loadtest.RequisitionFulfillmentServiceFlags
+import picocli.CommandLine
+
+class EdpSimulatorFlags {
+
+  /** The EdpSimulator pod's own tls certificates. */
+  @CommandLine.Mixin
+  lateinit var tlsFlags: TlsFlags
+    private set
+
+  @CommandLine.Option(
+    names = ["--data-provider-resource-name"],
+    description = ["The public API resource name of this data provider."],
+    required = true
+  )
+  lateinit var dataProviderResourceName: String
+    private set
+
+  @CommandLine.Option(
+    names = ["--data-provider-encryption-private-key-der-file"],
+    description = ["The EDP's Encryption private key (DER format) file."],
+    required = true
+  )
+  lateinit var edpEncryptionPrivateKeyDerFile: File
+    private set
+
+  @CommandLine.Option(
+    names = ["--data-provider-consent-signaling-private-key-der-file"],
+    description = ["The EDP's consent signaling private key (DER format) file."],
+    required = true
+  )
+  lateinit var edpCsPrivateKeyDerFile: File
+    private set
+
+  @CommandLine.Option(
+    names = ["--data-provider-consent-signaling-certificate-der-file"],
+    description = ["The EDP's consent signaling private key (DER format) file."],
+    required = true
+  )
+  lateinit var edpCsCertificateDerFile: File
+    private set
+
+  @CommandLine.Option(
+    names = ["--mc-resource-name"],
+    description = ["The public API resource name of the Measurement Consumer."],
+    required = true
+  )
+  lateinit var mcResourceName: String
+    private set
+
+  @CommandLine.Option(
+    names = ["--throttler-minimum-interval"],
+    description = ["Minimum throttle interval"],
+    defaultValue = "2s"
+  )
+  lateinit var throttlerMinimumInterval: Duration
+    private set
+
+  @set:CommandLine.Option(
+    names = ["--edp-sketch-reach"],
+    description = ["The reach for sketches generated by this EDP Simulator"],
+    defaultValue = "10000"
+  )
+  var edpSketchReach by Delegates.notNull<Int>()
+    private set
+
+  @set:CommandLine.Option(
+    names = ["--edp-sketch-universe-size"],
+    description = ["The size of the universe for sketches generated by this EDP Simulator"],
+    defaultValue = "10000"
+  )
+  var edpUniverseSize by Delegates.notNull<Int>()
+    private set
+
+  @CommandLine.Mixin
+  lateinit var kingdomPublicApiFlags: KingdomPublicApiFlags
+    private set
+
+  @CommandLine.Mixin
+  lateinit var requisitionFulfillmentServiceFlags: RequisitionFulfillmentServiceFlags
+    private set
+}
